requirements.txt
# Core dependencies
torch==2.5.1
numpy==2.1.2
scipy==1.15.3
pandas==2.3.3

# Simulation
simpy==4.1.1
networkx==3.3

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
pyyaml>=6.0
tensorboard>=2.14.0

# Optional (for notebooks)
jupyter>=1.0.0
